scopes consts moved to auth.go BroadcastUpdate now takes single scope UT updated
